requirements.txt
Authlib==0.15.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cryptography==3.4.8
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
